refactor : RepairShop 관련 리팩토링 진행 및 RepairShopMaker 추가
